scripts/requirements.txt
pandas>=2.2.0
python-dotenv>=1.0.0
supabase>=2.0.0
scikit-learn>=1.4.0
vaderSentiment>=3.3.2
scipy>=1.12.0
